Fix in IO OP queue for timeouts not occurring due to bug in recent changes made to fix out-of-order data
Fix in SMTP Client ReadAndSend & state change set that was causing issues with gmail, hotmail, aol & others due to continue sendfile being called after already done causing connection to hang until timeout:
"DEBUG" 1464 "2010-11-15 16:05:00.434" "SMTPClientConnection::_ParseASCII()"
"DEBUG" 1464 "2010-11-15 16:05:00.434" "SMTPClientConnection::~_BEFORE SendFile"
"DEBUG" 1464 "2010-11-15 16:05:00.434" "SMTPClientConnection::~_Continue sendfile"
"DEBUG" 1464 "2010-11-15 16:05:00.444" "SMTPClientConnection::~_AFTER SendFile"
"DEBUG" 2512 "2010-11-15 16:05:01.155" "SMTPClientConnection::~_Continue sendfile"
"DEBUG" 1464 "2010-11-15 16:10:11.862" "The read operation failed. Bytes transferred: 0 Remote IP: 74.125.95.27, Session: 3, Code: 995, Message: The I/O operation has been aborted because of either a thread exit or an application request"

Fix in SURL for URL's in plain-text emails where there was no trailing / causing very odd lookups. Debug logging added showed issue:
"DEBUG"	2300	"2010-11-21 15:15:35.982"	"SURBL:: Found URL: www.yahoo.com this is a test to see what happens when no "
"DEBUG"	2300	"2010-11-21 15:15:35.982"	"SURBL:: Lookup: yahoo.com this is a test to see what happens when no .multi.surbl.org"

Added user-customized min/max timeouts for SMTPC/SMTPD/POPC/POPD/SA/Clam (See New INI settings below) which help with premature timeouts when default is too low and locked box issues for iphone/cell users who lose connection before quit when default is too high.

Added ETRN server side: Upon delivery failure, if recipient is Route then message is put into HOLD state in queue. Upon successful ETRN command messages are released to be delivered again.

Added ETRN client side: For now no specific GUI area as "External Account" area is used temporarily. You put ETRN thehelddomain.com for Name, enter host/ip & 25 for port & minutes for how often. Rest of the settings are ignored.

Modified GUI Admin to show "Route Hold" for held ETRN messages.

Added option to replace sender IP in headers with fixed value such as 127.0.0.1 when auth'd as fix for barracuda & other blocking of email when end-user sender IP is blacklisted. IP is added to new X-AuthUserIP header instead. See INI setting below.

Changed SA so that results Temp file > original EML size to be considered not just > 0 which seems safer in event SA results are incomplete/corrupt. Since SA always adds SOMETHING (even just header) TMP should always be bigger than original otherwise just use original.

Added SA "move" file option vs copy+delete to help improve performance & cut down CPU/HDD load with debug logging to see which is in use & speed results.

Added new OnSMTPData(oClient) script event called after sender issues DATA which allows testing of sender/recipient(s)/IP & act without need to accept entire email 1st. 

Added new debug logging to help aid in tracking down cause of external account issues posted in forum. (suspect no header break..)

Added new POP server function to help filter TOP results (actual content of emails) from being logged. They are no longer logged unless LogLevel>8

Rejects for non-existent Route users now 421 temp vs 550 perm. Logic is it still stops secondary-relay spam but is safer if you define specific users in case list is not up-to-date. Should be as safe as greylisting since primary should accept eventually but could create a bit more traffic because of that.

Added new 3rd Result for OnAcceptMessage to allow for 453 temp fail result which is useful at times vs just 554 perm reject


New INI settings in hMailServer.ini [Settings]
ArchiveDir=C:\MailArchive
; Folder where copies of every email get stored. Default or undefined disables archiving
ArchiveHardLinks=1
; Enables use of NTFS Hardlinks. (NOT FAT/FAT32!) Default or 0 disables & forces normal Copy
POP3DMinTimeout=30
; 10 seconds is default which becomes 15 internally
POP3DMaxTimeout=90
; 600 seconds is default
SMTPDMinTimeout=90
; 10 seconds is default which becomes 15 internally
SMTP3DMaxTimeout=300
; 1800 seconds is default
POP3CMinTimeout=30
; 30 seconds is default
POP3CMaxTimeout=90
; 900 seconds is default
SMTPCMinTimeout=90
; 30 seconds is default
SMTPCMaxTimeout=300
; 600 seconds is default
SAMinTimeout=46
; 30 seconds is default
SAMaxTimeout=90
; 90 seconds is default
ClamMinTimeout=45
; 15 seconds is default
ClamMaxTimeout=90
; 90 seconds is default
SAMoveVsCopy=1
; Tells hmail to use move vs copy+delete. Defaults to 0 which means use "copy"
AuthUserReplacementIP=127.0.0.1
; Tells hmail to replace AUTH'd sender's IP with this value (to help when sender IP is on public blacklist). X-header is added instead. 
; Default is blank or use sender's real IP



git-svn-id: svn://hmailserver.com:13600@57 b719dfc6-fa5d-b644-ba71-0e501e651431

+++ trunk/source/Server/Common/AntiSpam/SpamAssassin/SpamAssassinClient.cpp
@@ -8,6 +8,8 @@
 #include "../../Util/File.h"
 #include "../../TCPIP/TCPConnection.h"
 
+#include "../../Application/TimeoutCalculator.h"
+
 #ifdef _DEBUG
 #define DEBUG_NEW new(_NORMAL_BLOCK, __FILE__, __LINE__)
 #define new DEBUG_NEW
@@ -17,7 +19,8 @@ namespace HM
 {
    SpamAssassinClient::SpamAssassinClient(const String &sFile) 
    {
-      SetTimeout(30); 
+      TimeoutCalculator calculator;
+      SetTimeout(calculator.Calculate(IniFileSettings::Instance()->GetSAMinTimeout(), IniFileSettings::Instance()->GetSAMaxTimeout()));
       
       m_sMessageFile = sFile;
    }
@@ -145,13 +148,26 @@ namespace HM
       bool bTestsRun = true;
 
       String sTempFile = m_pResult->GetName();
-      if (bTestsRun && FileUtilities::FileSize(sTempFile) > 0)
+      // In reality the SA response should always be bigger than the original message
+      // since SA should add SOMETHING to it, even if just a header or pre-pend subject
+      // So checking > orig EML makes more sense than just 0
+      if (bTestsRun && FileUtilities::FileSize(sTempFile) > FileUtilities::FileSize(m_sMessageFile))
       {
-         // Copy temp file to message file
-         FileUtilities::Copy(sTempFile, m_sMessageFile, false);
+         if (IniFileSettings::Instance()->GetSAMoveVsCopy())
+         {
+            // Move temp file overwriting message file
+            FileUtilities::Move(sTempFile, m_sMessageFile, true);
+            LOG_DEBUG("SA - Move used");
+         }
+         else
+         {
+            // Copy temp file to message file
+            FileUtilities::Copy(sTempFile, m_sMessageFile, false);
+            FileUtilities::DeleteFile(sTempFile);
+            LOG_DEBUG("SA - Copy+Delete used");
+         }
       }
 
-      FileUtilities::DeleteFile(sTempFile);
       return true;
    }
